Move equals demo to demos common
