--- demos/equals/pom.xml
@@ -1,47 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <parent>
-    <groupId>dk.alexandra.fresco</groupId>
-    <artifactId>demos</artifactId>
-    <version>1.2.2-SNAPSHOT</version>
-  </parent>
-
-  <modelVersion>4.0.0</modelVersion>
-  <artifactId>equals</artifactId>
-
-  <dependencies>
-    <dependency>
-      <groupId>dk.alexandra.fresco</groupId>
-      <artifactId>demos-common</artifactId>
-      <version>${project.version}</version>
-    </dependency>
-  </dependencies>
-  
-  <build>
-    <plugins>
-      <plugin>
-	      <artifactId>maven-assembly-plugin</artifactId>
-	      <executions>
-	        <execution>
-	          <configuration>
-	            <appendAssemblyId>false</appendAssemblyId>
-	            <archive>
-		            <manifest>
-		              <mainClass>dk.alexandra.fresco.demo.DistanceDemo</mainClass>
-		            </manifest>
-	            </archive>
-	            <descriptorRefs>
-		            <descriptorRef>jar-with-dependencies</descriptorRef>
-	            </descriptorRefs>
-	            <finalName>fresco-demo-distance</finalName>
-	          </configuration>
-	          <phase>package</phase>
-	          <goals>
-	            <goal>single</goal>
-	          </goals>
-	        </execution>
-	      </executions>
-      </plugin>
-    </plugins>
-  </build>
-</project>
